Move file to lib directory
